Reworks the color state system.
