+++ Controls/Components/ButtonComponents.cs
@@ -4,7 +4,8 @@ namespace UGUIDots.Controls {
 
     public enum ClickType : byte {
         PressDown = 1 << 0,
-        ReleaseUp = 1 << 1
+        ReleaseUp = 1 << 1,
+        Held      = 1 << 2
     }
 
     public struct ButtonClickType : IComponentData {
