+++ Core/Components/ButtonComponents.cs
@@ -1,25 +1,31 @@
+using System.Runtime.CompilerServices;
 using Unity.Entities;
 
 namespace UGUIDots {
 
-    /// <summary>
-    /// Defines if the state of the button based on the positions of the cursor(s).
-    /// </summary>
-    public enum ButtonState : byte {
-        None    = 1,
-        Hover   = 1 << 1,
-        Pressed = 1 << 2
+    public enum ButtonVisualState : byte {
+        None    = 0,
+        Hover   = 1 << 0,
+        Pressed = 1 << 1,
     }
 
     /// <summary>
     /// Stores if the button has a cursor on top of it, or if the cursor has been pressed 
     /// </summary>
-    public struct CursorState : IComponentData {
-        public ButtonState State;
-        // TODO: Deprecate the held...
-        public bool Held;
+    public struct ClickState : IComponentData {
+        public bool Value;
 
-        public static CursorState None() => new CursorState { State = ButtonState.None };
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static ClickState Default() => new ClickState { Value = false };
     }
 
+    /// <summary>
+    /// Stores the visual state of the button.
+    /// </summary>
+    public struct ButtonVisual : IComponentData {
+        public ButtonVisualState Value;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static ButtonVisual Default() => new ButtonVisual { Value = ButtonVisualState.None };
+    }
 }
