+++ Core/Systems/Render/UpdateLocalMeshDataSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1585ed336646da41bdd69304fa84355
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
