+++ Core/Systems/Render/UpdateMeshSliceSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf0e6bbbd1fcd5842ad6ae37444b52e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
